initial commit; using kube job approach, but abandoning that.
the idea of using GKE, having a node pool optimized for running the LH batch jobs, and
using kube Jobs to manage a bunch of LH runs in parallel seemed promising. the pool
could regularly sit at 0, then autoscale when the LH pods are scheduled, keeping the
cost low. but alas. the autoscaler is too slow for this to work the way i want.
testing with a job that manages 70 pods: the pods go PENDING, and the autoscaler
takes ~1m to add one node and have it become Ready; then another ~1m to add another; etc.

it took 5 minutes to run 70 pods this way, and i want to run ~5000, so this is a
show-stopper. something like https://github.com/atlassian/escalator seems
promising for this kind of rapid autoscaling of nodes for this kind of
bursty workload, but it only works on AWS.

keeping this commit here in case it is useful to someone or myself later.

gonna use AWS lambda instead for this project!